+++ polaris-core/src/test/java/org/apache/polaris/core/storage/azure/AzureCredentialsStorageIntegrationTest.java
@@ -22,7 +22,9 @@ package org.apache.polaris.core.storage.azure;
 import static org.apache.polaris.core.storage.azure.AzureCredentialsStorageIntegration.toAccessConfig;
 
 import java.time.Instant;
+import java.util.Optional;
 import org.apache.polaris.core.storage.AccessConfig;
+import org.apache.polaris.core.storage.StorageAccessProperty;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 
@@ -32,20 +34,34 @@ public class AzureCredentialsStorageIntegrationTest {
   public void testAzureCredentialFormatting() {
     Instant expiresAt = Instant.ofEpochMilli(Long.MAX_VALUE);
 
-    AccessConfig noSuffixResult = toAccessConfig("sasToken", "some_account", expiresAt);
+    AccessConfig noSuffixResult =
+        toAccessConfig("sasToken", "some_account", expiresAt, Optional.empty());
     Assertions.assertThat(noSuffixResult.credentials()).hasSize(2);
     Assertions.assertThat(noSuffixResult.credentials()).containsKey("adls.sas-token.some_account");
+    Assertions.assertThat(noSuffixResult.credentials())
+        .doesNotContainKey(
+            StorageAccessProperty.AZURE_REFRESH_CREDENTIALS_ENDPOINT.getPropertyName());
 
     AccessConfig adlsSuffixResult =
-        toAccessConfig("sasToken", "some_account." + AzureLocation.ADLS_ENDPOINT, expiresAt);
+        toAccessConfig(
+            "sasToken",
+            "some_account." + AzureLocation.ADLS_ENDPOINT,
+            expiresAt,
+            Optional.of("endpoint/credentials"));
     Assertions.assertThat(adlsSuffixResult.credentials()).hasSize(3);
     Assertions.assertThat(adlsSuffixResult.credentials())
         .containsKey("adls.sas-token.some_account");
     Assertions.assertThat(adlsSuffixResult.credentials())
         .containsKey("adls.sas-token.some_account." + AzureLocation.ADLS_ENDPOINT);
 
+    Assertions.assertThat(adlsSuffixResult.extraProperties())
+        .containsEntry(
+            StorageAccessProperty.AZURE_REFRESH_CREDENTIALS_ENDPOINT.getPropertyName(),
+            "endpoint/credentials");
+
     AccessConfig blobSuffixResult =
-        toAccessConfig("sasToken", "some_account." + AzureLocation.BLOB_ENDPOINT, expiresAt);
+        toAccessConfig(
+            "sasToken", "some_account." + AzureLocation.BLOB_ENDPOINT, expiresAt, Optional.empty());
     Assertions.assertThat(blobSuffixResult.credentials()).hasSize(3);
     Assertions.assertThat(blobSuffixResult.credentials())
         .containsKey("adls.sas-token.some_account");