+++ polaris-core/src/test/java/org/apache/polaris/core/storage/cache/StorageCredentialCacheTest.java
@@ -25,6 +25,7 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 import java.util.Set;
 import org.apache.iceberg.exceptions.UnprocessableEntityException;
 import org.apache.polaris.core.PolarisCallContext;
@@ -89,7 +90,8 @@ public class StorageCredentialCacheTest {
                 Mockito.any(),
                 Mockito.anyBoolean(),
                 Mockito.anySet(),
-                Mockito.anySet()))
+                Mockito.anySet(),
+                Mockito.any()))
         .thenReturn(badResult);
     PolarisEntity polarisEntity =
         new PolarisEntity(
@@ -103,7 +105,8 @@ public class StorageCredentialCacheTest {
                     polarisEntity,
                     true,
                     Set.of("s3://bucket1/path"),
-                    Set.of("s3://bucket3/path")))
+                    Set.of("s3://bucket3/path"),
+                    Optional.empty()))
         .isInstanceOf(UnprocessableEntityException.class)
         .hasMessage("Failed to get subscoped credentials: extra_error_info");
   }
@@ -121,7 +124,8 @@ public class StorageCredentialCacheTest {
                 Mockito.any(),
                 Mockito.anyBoolean(),
                 Mockito.anySet(),
-                Mockito.anySet()))
+                Mockito.anySet(),
+                Mockito.any()))
         .thenReturn(mockedScopedCreds.get(0))
         .thenReturn(mockedScopedCreds.get(1))
         .thenReturn(mockedScopedCreds.get(1));
@@ -137,7 +141,8 @@ public class StorageCredentialCacheTest {
         polarisEntity,
         true,
         Set.of("s3://bucket1/path", "s3://bucket2/path"),
-        Set.of("s3://bucket3/path", "s3://bucket4/path"));
+        Set.of("s3://bucket3/path", "s3://bucket4/path"),
+        Optional.empty());
     Assertions.assertThat(storageCredentialCache.getEstimatedSize()).isEqualTo(1);
 
     // subscope for the same entity and same allowed locations, will hit the cache
@@ -147,7 +152,8 @@ public class StorageCredentialCacheTest {
         polarisEntity,
         true,
         Set.of("s3://bucket1/path", "s3://bucket2/path"),
-        Set.of("s3://bucket3/path", "s3://bucket4/path"));
+        Set.of("s3://bucket3/path", "s3://bucket4/path"),
+        Optional.empty());
     Assertions.assertThat(storageCredentialCache.getEstimatedSize()).isEqualTo(1);
   }
 
@@ -164,7 +170,8 @@ public class StorageCredentialCacheTest {
                 Mockito.any(),
                 Mockito.anyBoolean(),
                 Mockito.anySet(),
-                Mockito.anySet()))
+                Mockito.anySet(),
+                Mockito.any()))
         .thenReturn(mockedScopedCreds.get(0))
         .thenReturn(mockedScopedCreds.get(1))
         .thenReturn(mockedScopedCreds.get(2));
@@ -178,7 +185,8 @@ public class StorageCredentialCacheTest {
             polarisEntity,
             true,
             Set.of("s3://bucket1/path", "s3://bucket2/path"),
-            Set.of("s3://bucket/path"));
+            Set.of("s3://bucket/path"),
+            Optional.empty());
 
     // the entry will be evicted immediately because the token is expired
     storageCredentialCache.getOrGenerateSubScopeCreds(
@@ -187,7 +195,8 @@ public class StorageCredentialCacheTest {
         polarisEntity,
         true,
         Set.of("s3://bucket1/path", "s3://bucket2/path"),
-        Set.of("s3://bucket/path"));
+        Set.of("s3://bucket/path"),
+        Optional.empty());
     Assertions.assertThat(storageCredentialCache.getIfPresent(cacheKey)).isNull();
 
     storageCredentialCache.getOrGenerateSubScopeCreds(
@@ -196,7 +205,8 @@ public class StorageCredentialCacheTest {
         polarisEntity,
         true,
         Set.of("s3://bucket1/path", "s3://bucket2/path"),
-        Set.of("s3://bucket/path"));
+        Set.of("s3://bucket/path"),
+        Optional.empty());
     Assertions.assertThat(storageCredentialCache.getIfPresent(cacheKey)).isNull();
 
     storageCredentialCache.getOrGenerateSubScopeCreds(
@@ -205,7 +215,8 @@ public class StorageCredentialCacheTest {
         polarisEntity,
         true,
         Set.of("s3://bucket1/path", "s3://bucket2/path"),
-        Set.of("s3://bucket/path"));
+        Set.of("s3://bucket/path"),
+        Optional.empty());
     Assertions.assertThat(storageCredentialCache.getIfPresent(cacheKey)).isNull();
   }
 
@@ -222,7 +233,8 @@ public class StorageCredentialCacheTest {
                 Mockito.any(),
                 Mockito.anyBoolean(),
                 Mockito.anySet(),
-                Mockito.anySet()))
+                Mockito.anySet(),
+                Mockito.any()))
         .thenReturn(mockedScopedCreds.get(0))
         .thenReturn(mockedScopedCreds.get(1))
         .thenReturn(mockedScopedCreds.get(2));
@@ -236,7 +248,8 @@ public class StorageCredentialCacheTest {
           entity,
           true,
           Set.of("s3://bucket1/path", "s3://bucket2/path"),
-          Set.of("s3://bucket/path"));
+          Set.of("s3://bucket/path"),
+          Optional.empty());
       Assertions.assertThat(storageCredentialCache.getEstimatedSize()).isEqualTo(++cacheSize);
     }
     // update the entity's storage config, since StorageConfig changed, cache will generate new
@@ -253,7 +266,8 @@ public class StorageCredentialCacheTest {
           PolarisEntity.of(updateEntity),
           /* allowedListAction= */ true,
           Set.of("s3://bucket1/path", "s3://bucket2/path"),
-          Set.of("s3://bucket/path"));
+          Set.of("s3://bucket/path"),
+          Optional.empty());
       Assertions.assertThat(storageCredentialCache.getEstimatedSize()).isEqualTo(++cacheSize);
     }
     // allowedListAction changed to different value FALSE, will generate new entry
@@ -264,7 +278,8 @@ public class StorageCredentialCacheTest {
           entity,
           /* allowedListAction= */ false,
           Set.of("s3://bucket1/path", "s3://bucket2/path"),
-          Set.of("s3://bucket/path"));
+          Set.of("s3://bucket/path"),
+          Optional.empty());
       Assertions.assertThat(storageCredentialCache.getEstimatedSize()).isEqualTo(++cacheSize);
     }
     // different allowedWriteLocations, will generate new entry
@@ -275,7 +290,8 @@ public class StorageCredentialCacheTest {
           entity,
           /* allowedListAction= */ false,
           Set.of("s3://bucket1/path", "s3://bucket2/path"),
-          Set.of("s3://differentbucket/path"));
+          Set.of("s3://differentbucket/path"),
+          Optional.empty());
       Assertions.assertThat(storageCredentialCache.getEstimatedSize()).isEqualTo(++cacheSize);
     }
     // different allowedReadLocations, will generate new try
@@ -291,7 +307,8 @@ public class StorageCredentialCacheTest {
           PolarisEntity.of(updateEntity),
           /* allowedListAction= */ false,
           Set.of("s3://differentbucket/path", "s3://bucket2/path"),
-          Set.of("s3://bucket/path"));
+          Set.of("s3://bucket/path"),
+          Optional.empty());
       Assertions.assertThat(storageCredentialCache.getEstimatedSize()).isEqualTo(++cacheSize);
     }
   }
@@ -310,7 +327,8 @@ public class StorageCredentialCacheTest {
                 Mockito.any(),
                 Mockito.anyBoolean(),
                 Mockito.anySet(),
-                Mockito.anySet()))
+                Mockito.anySet(),
+                Mockito.any()))
         .thenReturn(mockedScopedCreds.get(0))
         .thenReturn(mockedScopedCreds.get(1))
         .thenReturn(mockedScopedCreds.get(2));
@@ -322,7 +340,8 @@ public class StorageCredentialCacheTest {
           entity,
           true,
           Set.of("s3://bucket1/path", "s3://bucket2/path"),
-          Set.of("s3://bucket3/path", "s3://bucket4/path"));
+          Set.of("s3://bucket3/path", "s3://bucket4/path"),
+          Optional.empty());
     }
     Assertions.assertThat(storageCredentialCache.getEstimatedSize()).isEqualTo(entityList.size());
 
@@ -334,7 +353,8 @@ public class StorageCredentialCacheTest {
           new PolarisEntity(new PolarisBaseEntity.Builder(entity).id(1234).build()),
           true,
           Set.of("s3://bucket1/path", "s3://bucket2/path"),
-          Set.of("s3://bucket3/path", "s3://bucket4/path"));
+          Set.of("s3://bucket3/path", "s3://bucket4/path"),
+          Optional.empty());
       Assertions.assertThat(storageCredentialCache.getEstimatedSize()).isEqualTo(entityList.size());
     }
 
@@ -346,7 +366,8 @@ public class StorageCredentialCacheTest {
           new PolarisEntity(new PolarisBaseEntity.Builder(entity).entityVersion(5).build()),
           true,
           Set.of("s3://bucket1/path", "s3://bucket2/path"),
-          Set.of("s3://bucket3/path", "s3://bucket4/path"));
+          Set.of("s3://bucket3/path", "s3://bucket4/path"),
+          Optional.empty());
       Assertions.assertThat(storageCredentialCache.getEstimatedSize()).isEqualTo(entityList.size());
     }
     // order of the allowedReadLocations does not affect the cache
@@ -357,7 +378,8 @@ public class StorageCredentialCacheTest {
           new PolarisEntity(new PolarisBaseEntity.Builder(entity).entityVersion(5).build()),
           true,
           Set.of("s3://bucket2/path", "s3://bucket1/path"),
-          Set.of("s3://bucket3/path", "s3://bucket4/path"));
+          Set.of("s3://bucket3/path", "s3://bucket4/path"),
+          Optional.empty());
       Assertions.assertThat(storageCredentialCache.getEstimatedSize()).isEqualTo(entityList.size());
     }
 
@@ -369,7 +391,8 @@ public class StorageCredentialCacheTest {
           new PolarisEntity(new PolarisBaseEntity.Builder(entity).entityVersion(5).build()),
           true,
           Set.of("s3://bucket2/path", "s3://bucket1/path"),
-          Set.of("s3://bucket4/path", "s3://bucket3/path"));
+          Set.of("s3://bucket4/path", "s3://bucket3/path"),
+          Optional.empty());
       Assertions.assertThat(storageCredentialCache.getEstimatedSize()).isEqualTo(entityList.size());
     }
   }
@@ -451,7 +474,8 @@ public class StorageCredentialCacheTest {
                 Mockito.any(),
                 Mockito.anyBoolean(),
                 Mockito.anySet(),
-                Mockito.anySet()))
+                Mockito.anySet(),
+                Mockito.any()))
         .thenReturn(properties);
     List<PolarisEntity> entityList = getPolarisEntities();
 
@@ -462,7 +486,8 @@ public class StorageCredentialCacheTest {
             entityList.get(0),
             true,
             Set.of("s3://bucket1/path", "s3://bucket2/path"),
-            Set.of("s3://bucket3/path", "s3://bucket4/path"));
+            Set.of("s3://bucket3/path", "s3://bucket4/path"),
+            Optional.empty());
     Assertions.assertThat(config.credentials())
         .containsExactly(Map.entry("s3.secret-access-key", "super-secret-123"));
     Assertions.assertThat(config.extraProperties())