+++ runtime/service/src/main/java/org/apache/polaris/service/metrics/PolarisValueExpressionResolver.java
@@ -23,6 +23,7 @@ import io.micrometer.common.lang.Nullable;
 import jakarta.annotation.Nonnull;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
+import jakarta.ws.rs.core.SecurityContext;
 import org.apache.polaris.core.context.RealmContext;
 
 @ApplicationScoped
@@ -38,6 +39,13 @@ public class PolarisValueExpressionResolver implements ValueExpressionResolver {
         && parameter instanceof RealmContext realmContext
         && expression.equals("realmIdentifier")) {
       return realmContext.getRealmIdentifier();
+    }
+
+    if (metricsConfiguration.userPrincipalTag().enableInApiMetrics()
+        && parameter instanceof SecurityContext securityContext
+        && expression.equals("userPrincipal")
+        && securityContext.getUserPrincipal() != null) {
+      return securityContext.getUserPrincipal().getName();
     }
     return null;
   }