+++ runtime/service/src/main/java/org/apache/polaris/service/task/TaskUtils.java
@@ -19,7 +19,8 @@
 package org.apache.polaris.service.task;
 
 import java.io.IOException;
-import org.apache.commons.codec.binary.Base64;
+import java.nio.charset.StandardCharsets;
+import java.util.Base64;
 import org.apache.iceberg.ManifestFile;
 import org.apache.iceberg.ManifestFiles;
 import org.apache.iceberg.exceptions.NotFoundException;
@@ -46,9 +47,20 @@ public class TaskUtils {
    */
   public static String encodeManifestFile(ManifestFile mf) {
     try {
-      return Base64.encodeBase64String(ManifestFiles.encode(mf));
+      byte[] encodedBytes = ManifestFiles.encode(mf);
+      return new String(Base64.getEncoder().encode(encodedBytes), StandardCharsets.UTF_8);
     } catch (IOException e) {
       throw new RuntimeException("Unable to encode binary data in memory", e);
+    }
+  }
+
+  public static ManifestFile decodeManifestFileData(String manifestFileData) {
+    try {
+      byte[] decodedBytes =
+          Base64.getDecoder().decode(manifestFileData.getBytes(StandardCharsets.UTF_8));
+      return ManifestFiles.decode(decodedBytes);
+    } catch (IOException e) {
+      throw new RuntimeException("Unable to decode base64 encoded manifest", e);
     }
   }
 }